+++ rq/queue.py
@@ -1,8 +1,9 @@
 import times
 from .connections import resolve_connection
 from .job import Job, Status
-from .exceptions import NoSuchJobError, UnpickleError, InvalidJobOperationError
-from .compat import total_ordering
+from .exceptions import (NoSuchJobError, UnpickleError,
+                         InvalidJobOperationError, DequeueTimeout)
+from .compat import total_ordering, string_types, as_text
 
 
 def get_failed_queue(connection=None):
@@ -26,8 +27,9 @@ class Queue(object):
         connection = resolve_connection(connection)
 
         def to_queue(queue_key):
-            return cls.from_queue_key(queue_key, connection=connection)
-        return map(to_queue, connection.keys('%s*' % prefix))
+            return cls.from_queue_key(as_text(queue_key),
+                                      connection=connection)
+        return list(map(to_queue, connection.keys('%s*' % prefix)))
 
     @classmethod
     def from_queue_key(cls, queue_key, connection=None):
@@ -63,29 +65,50 @@ class Queue(object):
         """Returns whether the current queue is empty."""
         return self.count == 0
 
+    def safe_fetch_job(self, job_id):
+        try:
+            job = Job.safe_fetch(job_id, connection=self.connection)
+        except NoSuchJobError:
+            self.remove(job_id)
+            return None
+        except UnpickleError:
+            return None
+        return job
+
+    def get_job_ids(self, offset=0, length=-1):
+        """Returns a slice of job IDs in the queue."""
+        start = offset
+        if length >= 0:
+            end = offset + (length - 1)
+        else:
+            end = length
+        return [as_text(job_id) for job_id in
+                self.connection.lrange(self.key, start, end)]
+
+    def get_jobs(self, offset=0, length=-1):
+        """Returns a slice of jobs in the queue."""
+        job_ids = self.get_job_ids(offset, length)
+        return compact([self.safe_fetch_job(job_id) for job_id in job_ids])
+
     @property
     def job_ids(self):
         """Returns a list of all job IDS in the queue."""
-        return self.connection.lrange(self.key, 0, -1)
+        return self.get_job_ids()
 
     @property
     def jobs(self):
         """Returns a list of all (valid) jobs in the queue."""
-        def safe_fetch(job_id):
-            try:
-                job = Job.fetch(job_id, connection=self.connection)
-            except NoSuchJobError:
-                return None
-            except UnpickleError:
-                return None
-            return job
-
-        return compact([safe_fetch(job_id) for job_id in self.job_ids])
+        return self.get_jobs()
 
     @property
     def count(self):
         """Returns a count of all messages in the queue."""
         return self.connection.llen(self.key)
+
+    def remove(self, job_or_id):
+        """Removes Job from queue, accepts either a Job instance or ID."""
+        job_id = job_or_id.id if isinstance(job_or_id, Job) else job_or_id
+        return self.connection._lrem(self.key, 0, job_id)
 
     def compact(self):
         """Removes all "dead" jobs from the queue by cycling through it, while
@@ -95,7 +118,7 @@ class Queue(object):
 
         self.connection.rename(self.key, COMPACT_QUEUE)
         while True:
-            job_id = self.connection.lpop(COMPACT_QUEUE)
+            job_id = as_text(self.connection.lpop(COMPACT_QUEUE))
             if job_id is None:
                 break
             if Job.exists(job_id, self.connection):
@@ -106,7 +129,8 @@ class Queue(object):
         """Pushes a job ID on the corresponding Redis queue."""
         self.connection.rpush(self.key, job_id)
 
-    def enqueue_call(self, func, args=None, kwargs=None, timeout=None, result_ttl=None): #noqa
+    def enqueue_call(self, func, args=None, kwargs=None, timeout=None,
+                     result_ttl=None, description=None):
         """Creates a job to represent the delayed function call and enqueues
         it.
 
@@ -115,7 +139,7 @@ class Queue(object):
         contain options for RQ itself.
         """
         timeout = timeout or self._default_timeout
-        job = Job.create(func, args, kwargs, connection=self.connection,
+        job = Job.create(func, args, kwargs, description=description, connection=self.connection,
                          result_ttl=result_ttl, status=Status.QUEUED)
         return self.enqueue_job(job, timeout=timeout)
 
@@ -133,23 +157,24 @@ class Queue(object):
         * A string, representing the location of a function (must be
           meaningful to the import context of the workers)
         """
-        if not isinstance(f, basestring) and f.__module__ == '__main__':
-            raise ValueError(
-                    'Functions from the __main__ module cannot be processed '
-                    'by workers.')
+        if not isinstance(f, string_types) and f.__module__ == '__main__':
+            raise ValueError('Functions from the __main__ module cannot be processed '
+                             'by workers.')
 
         # Detect explicit invocations, i.e. of the form:
         #     q.enqueue(foo, args=(1, 2), kwargs={'a': 1}, timeout=30)
         timeout = None
+        description = None
         result_ttl = None
         if 'args' in kwargs or 'kwargs' in kwargs:
             assert args == (), 'Extra positional arguments cannot be used when using explicit args and kwargs.'  # noqa
             timeout = kwargs.pop('timeout', None)
+            description = kwargs.pop('description', None)
             args = kwargs.pop('args', None)
             result_ttl = kwargs.pop('result_ttl', None)
             kwargs = kwargs.pop('kwargs', None)
 
-        return self.enqueue_call(func=f, args=args, kwargs=kwargs,
+        return self.enqueue_call(func=f, args=args, kwargs=kwargs, description=description,
                                  timeout=timeout, result_ttl=result_ttl)
 
     def enqueue_job(self, job, timeout=None, set_meta_data=True):
@@ -183,10 +208,10 @@ class Queue(object):
 
     def pop_job_id(self):
         """Pops a given job ID from this Redis queue."""
-        return self.connection.lpop(self.key)
+        return as_text(self.connection.lpop(self.key))
 
     @classmethod
-    def lpop(cls, queue_keys, blocking, connection=None):
+    def lpop(cls, queue_keys, timeout, connection=None):
         """Helper method.  Intermediate method to abstract away from some
         Redis API details, where LPOP accepts only a single key, whereas BLPOP
         accepts multiple.  So if we want the non-blocking LPOP, we need to
@@ -194,12 +219,21 @@ class Queue(object):
 
         Until Redis receives a specific method for this, we'll have to wrap it
         this way.
+
+        The timeout parameter is interpreted as follows:
+            None - non-blocking (return immediately)
+             > 0 - maximum number of seconds to block
         """
         connection = resolve_connection(connection)
-        if blocking:
-            queue_key, job_id = connection.blpop(queue_keys)
+        if timeout is not None:  # blocking variant
+            if timeout == 0:
+                raise ValueError('RQ does not support indefinite timeouts. Please pick a timeout value > 0.')
+            result = connection.blpop(queue_keys, timeout)
+            if result is None:
+                raise DequeueTimeout(timeout, queue_keys)
+            queue_key, job_id = result
             return queue_key, job_id
-        else:
+        else:  # non-blocking variant
             for queue_key in queue_keys:
                 blob = connection.lpop(queue_key)
                 if blob is not None:
@@ -223,31 +257,35 @@ class Queue(object):
         except UnpickleError as e:
             # Attach queue information on the exception for improved error
             # reporting
+            e.job_id = job_id
             e.queue = self
             raise e
         return job
 
     @classmethod
-    def dequeue_any(cls, queues, blocking, connection=None):
+    def dequeue_any(cls, queues, timeout, connection=None):
         """Class method returning the Job instance at the front of the given
         set of Queues, where the order of the queues is important.
 
-        When all of the Queues are empty, depending on the `blocking` argument,
-        either blocks execution of this function until new messages arrive on
-        any of the queues, or returns None.
+        When all of the Queues are empty, depending on the `timeout` argument,
+        either blocks execution of this function for the duration of the
+        timeout or until new messages arrive on any of the queues, or returns
+        None.
+
+        See the documentation of cls.lpop for the interpretation of timeout.
         """
         queue_keys = [q.key for q in queues]
-        result = cls.lpop(queue_keys, blocking, connection=connection)
+        result = cls.lpop(queue_keys, timeout, connection=connection)
         if result is None:
             return None
-        queue_key, job_id = result
+        queue_key, job_id = map(as_text, result)
         queue = cls.from_queue_key(queue_key, connection=connection)
         try:
             job = Job.fetch(job_id, connection=connection)
         except NoSuchJobError:
             # Silently pass on jobs that don't exist (anymore),
             # and continue by reinvoking the same function recursively
-            return cls.dequeue_any(queues, blocking, connection=connection)
+            return cls.dequeue_any(queues, timeout, connection=connection)
         except UnpickleError as e:
             # Attach queue information on the exception for improved error
             # reporting
@@ -302,13 +340,14 @@ class FailedQueue(Queue):
             job = Job.fetch(job_id, connection=self.connection)
         except NoSuchJobError:
             # Silently ignore/remove this job and return (i.e. do nothing)
-            self.connection.lrem(self.key, job_id)
+            self.remove(job_id)
             return
 
         # Delete it from the failed queue (raise an error if that failed)
-        if self.connection.lrem(self.key, job.id) == 0:
+        if self.remove(job) == 0:
             raise InvalidJobOperationError('Cannot requeue non-failed jobs.')
 
+        job.status = Status.QUEUED
         job.exc_info = None
         q = Queue(job.origin, connection=self.connection)
         q.enqueue_job(job, timeout=job.timeout)