+++ rq/utils.py
@@ -5,8 +5,11 @@ Miscellaneous helper functions.
 The formatter for ANSI colored console output is heavily based on Pygments
 terminal colorizing code, originally by Georg Brandl.
 """
-import sys
 import os
+import sys
+import logging
+
+from .compat import is_python_version
 
 
 def gettermsize():
@@ -66,8 +69,11 @@ class _Colorizer(object):
         self.codes["darkyellow"] = self.codes["brown"]
         self.codes["fuscia"] = self.codes["fuchsia"]
         self.codes["white"] = self.codes["bold"]
-        self.notty = not sys.stdout.isatty()
 
+        if hasattr(sys.stdout, "isatty"):
+            self.notty = not sys.stdout.isatty()
+        else:
+            self.notty = True
 
     def reset_color(self):
         return self.codes["reset"]
@@ -121,3 +127,37 @@ def make_colorizer(color):
     def inner(text):
         return colorizer.colorize(color, text)
     return inner
+
+
+class ColorizingStreamHandler(logging.StreamHandler):
+
+    levels = {
+        logging.WARNING: make_colorizer('darkyellow'),
+        logging.ERROR: make_colorizer('darkred'),
+        logging.CRITICAL: make_colorizer('darkred'),
+    }
+
+    def __init__(self, exclude=None, *args, **kwargs):
+        self.exclude = exclude
+        if is_python_version((2,6)):
+            logging.StreamHandler.__init__(self, *args, **kwargs)
+        else:
+            super(ColorizingStreamHandler, self).__init__(*args, **kwargs)
+
+    @property
+    def is_tty(self):
+        isatty = getattr(self.stream, 'isatty', None)
+        return isatty and isatty()
+
+    def format(self, record):
+        message = logging.StreamHandler.format(self, record)
+        if self.is_tty:
+            colorize = self.levels.get(record.levelno, lambda x: x)
+
+            # Don't colorize any traceback
+            parts = message.split('\n', 1)
+            parts[0] = " ".join([parts[0].split(" ", 1)[0], colorize(parts[0].split(" ", 1)[1])])
+
+            message = '\n'.join(parts)
+
+        return message